Ex6: paddle.update() causes paddle.display() not to work
Problem: x and y velocity declared as xv and xy inside constructor function but then they're called vx and vy in the rest of paddle.js. This was causing paddle.update() to break.
Fix: replaced xv and yv at lines 13 and 14 with vx and vy respectively.
New lines: 14 and 16 (paddle.js)